backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
chromadb==0.5.23
openai>=1.0.0
pydantic==2.5.3
python-dotenv==1.0.0
